All helper methods moved into utils
