+++ test/utils.js
@@ -1,4 +1,8 @@
+const utils = require("@gnosis.pm/safe-contracts/test/utils/general")
 const BN = require("bn.js")
+const GnosisSafe = artifacts.require("./GnosisSafe.sol")
+
+const ADDRESS_0 = "0x0000000000000000000000000000000000000000"
 
 const jsonrpc = "2.0"
 const id = 0
@@ -29,7 +33,87 @@ function toETH(value) {
   return new BN(value * GWEI).mul(new BN(GWEI))
 }
 
+const execTransaction = async function(safe, lightWallet, to, value, data, operation, message) {
+  const nonce = await safe.nonce()
+  const transactionHash = await safe.getTransactionHash(to, value, data, operation, 0, 0, 0, ADDRESS_0, ADDRESS_0, nonce)
+  const sigs = utils.signTransaction(lightWallet, [lightWallet.accounts[0], lightWallet.accounts[1]], transactionHash)
+  utils.logGasUsage(
+    "execTransaction " + message,
+    await safe.execTransaction(to, value, data, operation, 0, 0, 0, ADDRESS_0, ADDRESS_0, sigs)
+  )
+}
+
+const execTransactionData = async function(gnosisSafeMasterCopy, owner, to, value, data, operation = 0) {
+  const sigs =
+    "0x" +
+    "000000000000000000000000" +
+    owner.replace("0x", "") +
+    "0000000000000000000000000000000000000000000000000000000000000000" +
+    "01"
+  return await gnosisSafeMasterCopy.contract.methods
+    .execTransaction(to, value, data, operation, 0, 0, 0, ADDRESS_0, ADDRESS_0, sigs)
+    .encodeABI()
+}
+
+const deploySafe = async function(gnosisSafeMasterCopy, proxyFactory, owners, threshold) {
+  const initData = await gnosisSafeMasterCopy.contract.methods
+    .setup(owners, threshold, ADDRESS_0, "0x", ADDRESS_0, ADDRESS_0, 0, ADDRESS_0)
+    .encodeABI()
+  return await getParamFromTxEvent(
+    await proxyFactory.createProxy(gnosisSafeMasterCopy.address, initData),
+    "ProxyCreation",
+    "proxy",
+    proxyFactory.address,
+    GnosisSafe,
+    "create Gnosis Safe"
+  )
+}
+
+const encodeMultiSend = async function(multiSend, txs) {
+  return await multiSend.contract.methods
+    .multiSend(
+      `0x${txs
+        .map(tx =>
+          [
+            web3.eth.abi.encodeParameter("uint8", tx.operation).slice(-2),
+            web3.eth.abi.encodeParameter("address", tx.to).slice(-40),
+            web3.eth.abi.encodeParameter("uint256", tx.value).slice(-64),
+            web3.eth.abi.encodeParameter("uint256", web3.utils.hexToBytes(tx.data).length).slice(-64),
+            tx.data.replace(/^0x/, ""),
+          ].join("")
+        )
+        .join("")}`
+    )
+    .encodeABI()
+}
+
+  // Need some small adjustments to default implementation for web3js 1.x
+  async function getParamFromTxEvent(transaction, eventName, paramName, contract, contractFactory, subject) {
+    assert.isObject(transaction)
+    if (subject != null) {
+      utils.logGasUsage(subject, transaction)
+    }
+    let logs = transaction.logs
+    if (eventName != null) {
+      logs = logs.filter(l => l.event === eventName && l.address === contract)
+    }
+    assert.equal(logs.length, 1, "too many logs found!")
+    const param = logs[0].args[paramName]
+    if (contractFactory != null) {
+      // Adjustment: add await
+      const contract = await contractFactory.at(param)
+      assert.isObject(contract, `getting ${paramName} failed for ${param}`)
+      return contract
+    } else {
+      return param
+    }
+  }
+
 module.exports = {
   waitForNSeconds,
-  toETH
+  toETH,
+  execTransaction,
+  execTransactionData,
+  deploySafe,
+  encodeMultiSend,
 }
